Fixed parent null bug with shared folders
Fixed a bug where a NullPointerException occurs if you owned a file, but the file is stored in someone else's Google Drive folder that is shared with you.

Now generates an ERROR in the log indicating that the parents could not be found.  The file may still be downloaded but will likely end up in the root path.

Also updated logging to not use AsyncAppender so any unexpected runtime exceptions (i.e. NPE) will still have chance of being logged instead of possible loss of unflushed log entries.